+++ app/src/main/java/ru/intelligency/scholarship/presentation/ui/portfolio/viewmodels/DocumentDetailsViewModelFactory.kt
@@ -6,7 +6,7 @@ import ru.intelligency.scholarship.domain.portfolio.PortfolioInteractor
 
 class DocumentDetailsViewModelFactory(
     private val interactor: PortfolioInteractor,
-    private val id: Long
+    private val id: Int
 ) : ViewModelProvider.Factory {
 
     @Suppress("UNCHECKED_CAST")
